2022 Day 16
LEADERBOARD BABY!

This was a rather nice surprise, after yesterdays failures! I was not
expecting to place today, as I felt rather slow to finish both tasks,
but it seems I did rather well, this being my highest rating on part 1
ever.

The task itself was the first challenging one this year. The first part
can be solved by simply using memoization to speed things up. The first,
and rather obvious, optimization here is to realize that you can
calculate how much pressure will be released by a valve, when you open
it, so there is no need to keep track of how much pressure is being
released at each time step.

The second is that there is no point in opening valves with a flow rate
of 0.

These two where enough to get part 1 across the line, but part 2 was
harder, as seen by the time. My first attempt was to simply adapt my
solution to part 1. This involved adding another parameter, and took a
quite a bit of fiddling to keep right, but I managed it. The only
problem was that it was far to slow. I applied more optimizations, such
as stopping once all valves where open, when that didn't work I stopped
when all valves with a rate higher than 0 was open. When that didn't
work I realized that it doesn't matter which player is where, and so
that reduced the search space by half.

I was able to calculate the answer for 23 minutes using this method, but
26 was going to take more RAM and time than I had. I started thinking
about ways to cut out states. One of them was to drop the nodes with a
flow rate of 0. To do this I would have to introduce the idea that the
path to different nodes had different cost. I didn't end up implementing
it in that form, but instead decided to pivot completely to a new
approach:

We being by building up the shortest paths between all the nodes. There
are so few, that this is very fast. We then build a set of the nodes
with a rate higher than 0, which are not yet opened. With these two
facts, we no longer talk about moving from room to room, but rather
which valve we are going to open next. This works, because there is no
benefit in moving, unless you are going to open a valve, so we can just
look at the total cost of opening a particular valve.

Doing it with 2 actors turned out to be a bit tricker, I tried to
integrate it into the system I already had for the older model, but they
differed to much. In the end, I introduced two separate remaining times
for the two actors, and simply simulated each of them moving to a given
valve, opening it and so forth. As long as the lists of valves are
shared, so they are not trying to open the same valve, they don't have
to think about each others timestep at all.

When I ran this on the example it returned very quickly, and I was
confident that I had now found the solution. I let it run, and became
more and more nervous, because it took quite some time, but in the end
it produced and answer, and it was correct.

In hindsight, I should have pivoted to this new method before, I spent a
bit too long trying to bruteforce this. I don't feel to bad about this
however, because it could just have easily have worked, and paid off
handsomely.

Seeing as both of my solutions here are rather slow, I really look
forward to optimizing this one!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 16   00:16:11     27     74   01:22:08    192      0