2022 Day 18
Well, that was kind of a fun day. Not much to say here, other than that
I spent way too long understanding the problem text in part 2. I should
really have had a function for orthogonally connected cubes, but I only
have it for squares. It might have given me a better time on part 1.

The part two which you see in the repo, is a reconstruction from memory,
as I managed to delete the real one. It has not been cleaned up however,
but rather I've tried to rewrite it as I had it when I submitted it.

I was anxious that part two was going to be really hard, but a simple
DFS search from outside the lava droplet is enough to solve the task
very quickly. While rewriting part two from memory, I hit onto all sorts
of small bugs in the DFS part, but luckily I did not hit any of these
during the actual contest.

I'm still not done with 17 part two, and I don't really have the first
idea on how to do it, but those are the best tasks ;)

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 18   00:05:18    472      0   00:22:46    583      0