2022 Day 20
Well, this is a big mess!

The task itself is cool. There are multiple of the same number, so we
need to keep track of them individually. I think I could have just used
an array for this, but I ended up using the awesome bidict module, to
make sure I kept everything in order.

I really tried to do this more optimally than this, moving multiple
things at once, instead of doing one iteration at a time, but I just
couldn't get the indexing right. There is some small bug there, that
cost me atleast 1 hour. I decided to revert to brute force after a while
for part one, and with the normal python interpreter it took something
like 10 minutes. This made part two seem infeasible to do the same, but
I actually ended up getting it this way, because pypy  was actually fast
enough to make it viable.

I didn't realize this before it had gone later though, I was actually
just working on the bruteforce method again, to create something to
compare the real algo to, but then I noticed that once i figured out how
to reduce the jumps, I could actually use it with pypy.

One tricky thing, is that when you consider us going a full round, that
is actually just N-1 moves. So if we have 3 numbers, you only need 2
moves to the left, to be back where you started. I realized this while
setting up the brute forcer for part two, maybe this is what was wrong
with the real algo. I'll go back and check that!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 20   01:30:37   2424      0   02:02:29   2507      0