2022 Day 17
Well, this was a really nice task! I did decently in part 1, coming in
at 140, but part two didn't come easy.

As with all these tasks, my first idea was to use dynamic programming.
This turns out to be the right move, but there was one crucial flaw,
which was that the example does not build floors. That was how I was
hoping to do the dynamic programming, by recognizing when a floor had
been made.

I tried various other approaches, but then after 1 and a half hour I had
to call it quits. I looked at it today and this time I tested it on my
actual input. Right away it was able to find a floor. With this
realization, I knew that my original idea was going to work.

It was still a rather long road until I got it implemented, mostly due
to me being slightly daft. Multiple small bugs later, I was able to make
a proper dynamic programming solution.

I then realized that it was not fast enough, but that once we had all
the positions figured out, we could simply see how big the loop was and
skip a lot of iterations. Then I got the proper result.

I am very curious as to how to solve the example though, there must be
some sort of "floor" function that will work for that too.

This one will be not nice to clean up, but I hope to be able to soon!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 17   00:26:26    140      0       >24h  13817      0