2022 Day 24
It's really weird, doing these tasks at a reasonable time like 12:00.
I'm so used to getting up early to do these tasks, that it was quite a
change ;)

I did OK on today's task, but I could have done a bit better. I spent
quite a bit on time debugging a one off error on the blizzards on part
1, but once I figured that out, it was not so hard.

Initially I began doing this through a BFS search, but this took too
much time, so I switched to A* with the heuristic being the Manhattan
distance to the goal. Initially I thought this might be to slow too, but
with pypy it was actually able to finish. This is yet another proof that
it never hurts to let the bruteforce solution running in the background,
given that it doesn't take up to much memory.

One smart thing I did early, is that I created a step function for the
blizzards, and I cached the output of this function. This allowed the
code to flow very naturally, but I didn't have to waste time
recalculating a bunch of conditions every timestep. It also brought
memory usage down, as I was able to reuse the frozensets from before.

It could have been done better, by pre calculating the answer for the
period of the system, which is the width of the map - 2, and then just
referring to a certain setup, by the timestamp modulo the map width. But
this could have been error prone, and my solution was easier to reason
about.

For part two, I just used the same search 3 times. I was initially
worried that I might need to consider the route as a whole, as opposed
to just focusing on each leg, but this turns out not to be the case.
Since you have the option of waiting, there would never be a case where
we arrive later at the destination, and have a "better" blizzard
configuration than just doing the fastest route and then waiting. This
is because the blizzards are locked to the time step.

I only realized this while the program was running, but I started the
search on the assumption that it might be correct. This is another case
where making assumptions about the task leads to the correct solution.

My current solution is very slow, and I'd like to look into how to make
it faster. I haven't really been cleaning up the tasks this year, so it
will be nice to go over them again and see what can be improved.

Now I just need to finish part 2 on day 22, and I'm ready for tomorrow!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 24   00:40:18    537      0   00:48:46    555      0