2022 Day 22
This one was relatively painless, except for the fact that I made one
mistake on part two, that made it so that I had to do it later.

I had time pressure when solving this, as we where taking a plane to
Thailand that morning. So I only gave part 2 one try and then just
closed the program.

Part 1 is very easy, and I scored quite well on that. For part two I
thought it was going to be a massive headache to solve generically, and
I still think it will be, but I ended up going for something which was
just for my input. I decided on how to name the sides on my input and
drew it out on a piece of paper. Then I cut the paper to that shape, and
could fold it like a cube.

I had written the letter for each side on the cubes on the paper, so I
could track how their rotation ended up. Then it was an easy matter to
figure out which face I was currently on, and if a move would end up in
another face.

This is where things get a bit interesting. The transformation between
two faces is not trivial. But I drew it out on the paper and wrote a
transformation setup by hand. I decided I would have (new face,
direction entered on new face, reverse numbering, new direction).

This worked ALMOST perfectly, but with one mistake, which is that I had
ONE mistake in the whole thing, a single "L" that should have been an
"R" in the "new direction" column. The most annoying part, is that the
"new direction" is the only part of the solution, that is not needed at
all. It should always be the opposite of the side we enter. So I could
have just not hardcoded it and I would have been finished on time.

Either way, I'm now set for tomorrow. This is very nice as I was worried
that the mistake was going to be harder to track down. I can't wait to
see how other people have solved this in the reddit thread, but I will
first try to solve it generally on my own before looking.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 22   00:30:35    390      0       >24h   9323      0