2022 Day 23
Well, we finally got the mandatory "Conway's game of life" variation
puzzle. This one was fairly straight forward, except that I wasted A LOT
of time on part one, on a shitty bug in my "adjacency" function. I
forgot to exclude the "(0,0)" delta, and so I got 9 items, instead of
just the 8 squares surrounding.

I had to create a visualisation of this to figure it out and that took
some time. This is the 3rd time this year, that I wished I had an
unconstrained adjacency function, so I think that I'll include one like
that in next year.

Part two is not very fast, even when using pypy. I wonder if it's me
doing something stupid or if it's just a slow task. I will look in the
megathread to see if there are some smart ways of optimizing this.

The reason I was not able to deliver this when the contest started, was
that I was on a plane, with no elbow room to really work. So I decided
to wait until we had arrived at our AirBnB.

Actual time:

Part 1: 00:39:23
Part 2: 00:01:46
Total : 00:41:09

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 23   04:25:17   4870      0   04:27:03   4640      0