2022 Day 25
Well, another advent of code done! I think there has been a couple of
good days this year, 17, 19 and 22 come to mind. Overall though it has
been rather tame, but this has been the trend throughout.

I was ok pleased with my solve of this, even though it is a bit
clumsy to convert from integers to snafu numbers using z3. I am sure
there is some easy way to do this, which is just to find the biggest
number needed to represent a given number and then iterating on that,
but I didn't want to risk it.

I had a small bug in the int to snafu code, which cost me a lot of time,
I was starting with a 1, instead of a 0. Once this was fixed though, it
ran relatively quickly.

I look forward to cleaning up some tasks this year, but it won't be
right away, as I don't feel like programming a lot when we are in
Thailand. This has been the first year, where someone from work has
given me a real run for my money. This should definitely be repeated!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 25   00:15:31    510      0   00:15:35    429      0