Cleanup of 2022 Day 15 Part 2
This turns out to have a beautiful solution! I looked at the megathread
to get this one, but it depends on just looking at the equations for the
borders around the sensors. This brings the number of positions we have
to consider drastically. You can further bring it down, by applying some
more restrictions. I've tried to write some notes about this in the
program itself.