2023 Day 01
I am writing this log retroactively, because I'm in Italy at the moment,
and it's quite hectic. I will try to be better about this in the future.

I have a couple of things going against me this year. Number one is that
I haven't practiced any, so there is bound to be some rust in the system
that will screw me over in the future. Number two is that I haven't
typed on this keyboard before, so there is bound to be some frustration
there too. This being said, I am very excited for the tasks this year!

The tasks today was ok, I spent a bit of time on understanding the first
part, but then it went by rather quickly. The problem came on the second
part, where I missed out on the important clue, that there can be
overlapping matches. After I figured this out, it was not so hard to
solve it. I am pleased with the times, especially as I am not expecting
much from this beginning!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  1   00:01:50    153      0   00:19:10   1624      0