2023 Day 3
Today we camped in a lovely place in the dolomites, but it had no
internet access. We slept in the car and I only got internet a couple of
hours later. So I timed myself, and the time is given below. It would
not be enough to get any good scores anyway. It does mean that my
perfect streak at work is broken though, which is a bit sad.

The task today was very interesting! I settled on a solution quite
quickly and given I had a preprogrammed `adj` function, it was easier
than it could have been. I wonder what the fast solution is here.

I am proud of how I pivoted my part one solution into the part two
solution. I hit only really one snag and that was not using a set
immediately, but rather a list. It produced duplicates. It's a clear
error in hindsight, but at the time it gave me some trouble.

Part one: 00:07:21.09
Part two: 00:12:59.00
Total:    00:20:20.80

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  3   05:57:00  28244      0   06:04:21  22241      0