2023 Day 4
Ok, so the tasks are getting more complicated, especially the parsing
stage. So mister Wastl is trying to make life hard for the AI
specialists.

Overall, I was very slow today. I made a couple of mistakes and had to
wait 60 seconds for both parts. I also had a couple of mistakes in
implementation which slowed me down.

The first implementation mistake was that I didn't remember that my
`numbers` list, went line by line, I thought it was all the number in
one long list. Since I didn't remember this, I tried to use the
`mit.chunked` function, which created some weird error messages. Very
obvious in hindsight, but I lost a lot of speed to it.

The big mistake I made on part one was not realizing that the punch
cards in my input was 10 long, instead of 5. After I realize this, I get
the right input.

The other thing, was that I was slow to code up a function for the value
function, aka 1, 2, 4, 8, 16. I realize now it should have been
`pow(2,n-1)`, but in the moment I ended up just coding it.

For the second, I had to resolve to pypy to get it to complete in a
reasonable time. I know a 100 this can be done via memorization, but I
just didn't want to code it up, unless it was necessary.

Coding this up will be nice, I can't wait to improve it.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  4   00:13:34   4916      0   00:21:58   2611      0