2023 Day 6
Well, this was very easy???

I don't know if it's just because it went easy for me, but this felt
like a day 2 or 3 task. Even part two could just be easily brute forced.

I know that in the end this is a very easy mathematical problem, which
I'll clean up later, but I thought that there was no point in going
there right away, as the input was not that big.

I also decided to just input the example and my input by hand, which in
hindsight was a mistake, as it was just a simple "zip" and string
concatenation in python, but I think I lost the most time here, to small
implementation quirks and just problem parsing.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  6   00:05:46    763      0   00:07:12    422      0