Cleanup of 2023 Day 4
This turned out extremely clean! I must admit here that I got the idea
of the optimization of this, from the youtuber "john paulsen" I think.
When I saw how he solved it so quickly, I was inspired. The trick is
very cool and allows for iteration based solution.

The trick is also enabled by the fact that we will not have any "extra
items" at the end, because the text promises us we will not be asked to
copy cards not in inventory.

It would be possible to make this not memory bound by doing accumulation
of the score as we process each card and also deleting them from the
default dict as we go. This would make the algorithm totally streamable.