Cleanup of 2023 Day 6
This was a rather nice cleanup. It turned out to be a quadratic equation
like I thought it would be. I struggled a bit with getting it right,
because in some instances it would hit the perfect roots, which is not
what I wanted.

I expect that my current solution might be inaccurate for certain
inputs, but it works for now :)