2023 Day 7
Another problem parsing heavy day. It was rather straight forward, but I
spent a lot of time on parsing the tasks.

For part two I spent a little time realizing that the optimal move was
to always replace the card with the second best one. I briefly
considered doing the permutations of Js, but I was lazy and didn't want
to do it.

I placed very badly today, but I'm not to mad about it. I can't wait to
do better tomorrow.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  7   00:20:18    991      0   00:33:15   1119      0