2023 Day 8
So I did this while I was hungover from an office party, so there is
that to consider. The first part went down easy, I spent a little longer
on part two.

I didn't really like this task. It's another example of finding the
solution via least common denominator, but it's all in the input. I
prefer that if the puzzle text allows for nasty situations, then the
input should allow for it. I guess that it's a good skill to have, to
play dirty and make assumptions based on your input, but it doesn't quit
sit right with me.

There fact that I was able to solve this by doing my solution hinges on
two things from the input:

- All loops are multiples of the length of the LR instructions. This
  means they will always be in sync.
- All loops have the same initial loop as they have repeat loop times.

I haven't looked at the task more, but it appears to me that without
these limitations in the input, it would be a much tougher task. I will
see if I can figure out a way to solve this, that doesn't rely on these
two assumptions.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  9   00:10:34   1393      0   00:13:17   1167      0
  8   00:05:05    419      0   00:21:48    879      0