Cleanup of 2023 Day 9
This ended up being ultra clean when I just cleaned it up a little. I
wonder if there isn't a way to avoid keeping all rows in memory. I know
we could just keep the last and first elements of everything except two
rows, but I don't see the point in optimizing for it, when the current
memory usage is so small.