2023 Day 10
Wooow! What a day!

I spent a bit to long parsing the problem on part 1, and I made a fatal
mistake of thinking I should use the "graphlib" topological sorter for
this, which ended up wasting even more time. I did however make some
good decisions along the way:

- I hard coded my value of S, so I didn't have to write any detection
  logic for it. This meant I had to keep in mind which example I was
  working with, but I believe that it saved time overall.
- I implemented BFS, with a simple depth counter very quickly, which was
  nice.
- I created aliases for the directions, which speed up writing a lot.

There is not much more to say about part 1, it went pretty good, it was
mostly reading comprehension and the graphlib misstep that got me.

Now part two was a different kind of saga. As seen by my score, it took
quite some time to resolve. There was multiple reasons for this, but
most of them stemmed from reading comprehension and small bugs.

My idea became after a little while to implement "virtual coordinates",
that is creating a new grid where there is space between all actual grid
tiles. Thus, if I could find a single ground tile that belong inside the
loop, I could use flood filling to find the rest. Afterwards, I could
just count up the "real coordinates" in the flood fill and use that.

This theory works, but implementing it turned out to be a real hellhole.
I also didn't realize until very late that pipe pieces not used in the
loop, but inside it, counts as open spaces. This lead to the most
annoying bug, which was that I had removed the "ground areas" around
these pipes, but now that they where open, I could traverse over them
again.

What I ended up doing, which was very cheeky, was to do a two step
algorithm. First I would "clean the board", by replacing all spaces not
included in the loop with ground tiles. Then I would run it again and
get the right answer. This is super janky, but it works.

I looked at the subreddit before writing this and I found a MUCH EASIER
way to do this, which is to do raycasting, line by line. It works
perfectly and is a fraction of the code. I'll clean it up and implement
it now. I am a bit annoyed that I didn't think of it here, but I don't
think I could have done it as elegantly.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 10   00:13:36    191      0   01:43:24   1980      0