2023 Day 12
Well well well well!

Today was challenging and I liked that. I was slow overall, it shows as
I spent a lot of time implementing just part one. I should have done it
simpler to begin with, but I wanted to be smart. In the end, just
iterating the possibilities was fast enough for part one.

The bigger issue came with part two as I guessed. I started trying out
different techniques, like pruning options early. I realized quite early
on that we can eliminate all strings of "." and blocks of "#" at the
start and end. I tried memorization on one function first, but it didn't
lead to big performance gains so I dropped the idea.

I continued trying to find new ways to prune the tree, but in the end it
got stuck on certain inputs. Look over the iterations again I saw a lot
of repetitions, so I tried adding memorization again as a last ditch
effort before actually trying to do something smart. I don't know if
it's because I structured the code differently or that one or more of
the pruning techniques was very good at reducing the input space, but
now it flew through the task.

I could have solved this about 30-40 minutes faster because of this I
think, I'll try to experiment to see which pruning techniques actually
make a difference.

Overall a really nice day and it shows that you should always add
memorization to any place you think it could be useful!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 12   00:29:45   1855      0   01:50:08   2076      0