Cleanup of 2023 Day 12
Ok, so this is my initial attempt to optimize this. The big change now
is that we are placing each block directly, so as to avoid invalid
permutations entirely. This still feels too slow for me, so I'll take a
look at reddit and see if there isn't another way to solve this.