2023 Day 17
Well, well, well, well, Dijkstra's algorithm makes it first appearance
of the year and it's actually a bit exciting this time!

I realized there would be some sort of graph solution here quite
quickly, but I spent a bit of time fiddling to implement the algorithm.
I had to google it as usual, I never remember the implementation
immediately.

I ended up thinking of the nodes as dependent on:

(position, traveling directions, steps traveled in this direction)

With these 3 dimensions, we are free to explore the full potential
range.

The algorithm would be to just for each node, I would generate a list of
possible next nodes, one being going further ahead the current
direction, if the number of steps where less than 3 and the rest with 0
steps, but the directions being left and right.

I made some small mistakes in implementation. One was not filtering out
the current direction in the "all other directions" step, I only
filtered out the opposite direction.

Another mistake was that I was allowing the bot to go 4 steps, not 3.

The other was that I set the destination as (Y,X) when it should have
been (Y-1,X-1) of course. This was the biggest time waste, as it forced
me to implement debug printing, with the "prev". This wasn't so bad
though, as it caught the two previous mistake actually.

For part two, I just had to add an "if" around the other directions
part, so we only added them after having walked 4 steps.

The current solution is slow, and there are some obvious improvements
that could be made, at least for part 2. I could skip the first 4 steps
and just jump immediately the minimum 4 steps. This would cut out a lot
of small steps.

A* could also potentially speed this up a bit, let's see!

The biggest improvement of course, would be to reduce the dimensions.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 17   00:36:31    903      0   00:39:12    644      0