2023 Day 18
So this ended up being harder than expected for part one and easier than
expected for part two.

For part one, I tried implementing a sort of scanline algorithm that I
had done for day 10, but in the end, something was wrong with my idea,
or something was wrong with my implementation, so I had to resort to
using BFS and filling the inside of the polygon to get the answer.

For part two, I realized this would be impossible and I was reminded of
some solution for day 10 that I saw on reddit, which was the "shoelace
formula" and "Pick's theorem". I was a bit confused on how to combine
them in the beginning, but I got it in the end.

I wasn't on the leaderboard for either one, but it ended up being a fun
day, though a bit of a skillcheck. Either you knew about this formula,
or you didn't and it wouldn't work out for you.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 18   00:25:29   1219      0   00:57:17   1067      0