2023 Day 19
Well, I didn't wake up for this one either, so I took it later in the
day. I would have been nowhere near the leaderboard either way, so it
was a good choice.

Part one was a bit kinky to implement, I got screwed over by some sort
of lambda capture properties in python, which didn't let me do what I
originally intended for the workflows. But I implemented it differently
and it went through. Only snag was that I didn't see that we always
started with the "in" node, but once I fixed that, it went through.

Part two was harder, but I realized that what I could do was split it up
into ranges and then at each step, subdivide them up and just look at
all of the ones that passed. This ended up working amazingly, the only
problem was that I entered in "40000" instead of "4000", and I spent a
lot of time looking at this, before I got the right answer.

One of the coolest days in the year, I like these tasks where you have
to think in ranges!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 19   07:02:22  12430      0   15:58:10  12037      0