2023 Day 23
This day was very interesting! It is not so often you try to find the
longest path, and using directed graphs like this is also not so usual.

Part one was simply implementing a dfs, with a bit of a change. I have
done this in previous years, but as we are more interested in not
repeating a step on the same path, we are keeping a list of all the
places we have been in the stack. The reason for using DFS, is that we
don't want the memory usage to explode.

I didn't convert the grid into a graph, I just read it directly off the
board. I don't usually do this, but I thought in this case, maybe it
would save some time.

For part two, I first just tried to bruteforce it, the same way as
before. This didn't end up working. I did a little sneaky tactic here.
AoC gives you like a hint if your input is way to low or way to high,
but it doesn't say anything once you actually enter into a certain range
around the answer. This is to avoid people binary searching the answers
I guess.

It does however, that you can figure out if you are in the neighborhood
of the answer. Once the incremental improvements that was printed out
reached the point where it didn't say if it was above or below, I knew
that if I could just speed up the bruteforce a little bit, then it would
work.

My idea was that I could collapse the graph, by recording the distance
between the different "intersections". There are no other places to go,
so we can just build a graph where the nodes are the the intersection
points.

The I just ran the same bruteforce algorithm that I did before and with
pypy it finds the answer very quickly. It still takes a time to iterate
through all options, I am sure there are some faster solutions out
there, I can't wait to look at the reddit thread to figure out what the
optimal way of doing this is.

Part one: 00:17:39
Part two: 00:51:39
Total   : 01:09:18

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 23   16:04:59  11141      0   16:56:37   7854      0