2023 Day 24
I didn't wake up for this one either, but at least I did it the same
day!

Today was an ultimate z3 day! I had no idea how to solve this normally,
but z3 showed the way on both parts. For part one I just tested if there
was a solution. This took quite some time, so it actually took most of
the time, just checking. I am not sure there is a way to speed this up,
but it worked.

Part two was even easier and solved much faster.

I have very very little idea of how to do this without z3, but I'll chew
on it for a bit, as it seems like a very good training task. My current
idea for part one, is to convert them over to normal:

y = m*x + b

Equations, to be able to use normal collision detection. We will see if
it works!

Part one: 00:29:40
Part two: 00:09:05
Total:    00:38:45

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 24   07:16:47   6992      0   07:25:53   2953      0