2023 Day 25
OK, well this is definitively the hardest final day of any year I've
participated in, but I got it in the end.

I began with an idea of just testing all possible permutations, but that
didn't pan out. I spent too much time just waiting on the brute force,
because I didn't really have any other ideas. Googling for the problem
didn't really reveal any ready to use algorithms.

So for the second time this year, I used graphviz to visualize the
example graph and I noticed that there was really two clusters,
separated by two bridges. I thought that if I could find two nodes that
I knew would be in different clusters, then the edges that I would have
to remove, would have to be in the shortest path between them. Once I
could no longer connect them, I would know that I had split the graph in
two.

Once I started implementing this, I simply created a iterative search to
see which pairs was the furthest apart. This was actually one of the
slower part of the task, ideally I would want the two items that was
furthest away from each other, as I would guaranteed that they where in
two different clusters, but it didn't need to be the ones with the most
space between them.

In the end I ended up printing the progress and then choosing a random
pair, once progress kind of stopped.

The idea then, is that we just compute the shortest path between these
two nodes, and see what that path contains. We consider all of these
edges to be potential removals, so we put on the stack that each of
these should be considered eliminated. If the set of eliminated nodes
exceeds 3, we just return nothing, but if there is no path between the
two nodes, then we know that we have split the groups in two.

This actually runs very quickly, with the precomputed A and B nodes.

I am interested to solve day 21 part two, as I haven't done it yet and
to see what I can learn on reddit about todays task!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 25   01:26:44   1931      0   01:26:48   1649      0