Add additional solution to 2023 Day 25
I keep reading about solutions to this one, from a fantastic reddit
comment:

https://old.reddit.com/r/adventofcode/comments/18qbsxs/2023_day_25_solutions/kftp4jr/

I am not getting any times even close to the times listed here, I don't
know if that is python or implementation slowness. Their code isn't
exactly super easy to read and I'm not so used to C#, so it will take a
while, but I'll try to tease it apart to see how it works.

I got to use a Disjoint-set data structure here as part of Kruskal's
algorithm. I remember finding this data structure for the first time in
another programming competition and being really impressed with it's
simplicity.

I'll continue to try to implement this correctly, before moving on to
task 21 part two, which I have still to implement