More updates of 2023 Day 25
To try to figure out why the karger stein implementation was so slow, I
turned to some profiling tools. It's been a while since I've dabbled in
python profiling, but a quick google search revealed two new tools to
me: py-spy and pyinstrument. Both revealed more about what is taking
time in my program.

For the DSU version, it showed that I might be able to eliminate some
overhead, by removing the DSUSet class, and instead storing two dicts
directly in the DSU class. This didn't give any big speedups, but it
improved the runtime a little.

For the karger stein algorithm I was able to determine that a huge
amount of time was spent in the "deepcopy" module! Searching around, I
now see that this is one of the slowest ways of deepcopying data in
python, because it has to support circular datastructures. If your data
doesn't have these type of cycles, then the "marshal" package seems to
be the quickest way to deepcopy data in python, if you are not prepared
to write your own loop, like I was. This sped the program up many times!

I am fairly certain that the remaining time is spent largely on the
"del" operations, but I am having a hard time measuring this as the
profiling tools I have, is not showing the delete operations.

I'm still interested in why my implementation seem to take so much
longer time than those of other people, but that will be a problem for
tomorrow!