Improve 2023 Day 24 part 1
This is a very interesting task and I've actually ran into some problems
implementing it correctly without z3.

The basic idea is that we need to express all the lines as functions of
the same variable, and then try to see if we can eliminate that
variable. The way I chose to do this, is to convert them all into
equations of the form

y = m*x + b

We can do this fairly easily, because the functions for x and y for the
hailstone paths are linear equations and so (fx(s), fy(s)) is also a
linear equation, which we can convert to the above form, by simply
evaluating it for 2 points and then using these to find the normal
formula.

Once we have them as a normal formula, we just have to figure out for
which X, their Ys are the same. There is either zero, infinite or
exactly one solution to this problem. There is actually a tricky part
here, which is that we need to then calculate the corresponding time
which this would have occurred for both lines and make sure it's
positive. If it's not, it's happened in the past.

When I initially implemented this,I actually got too many positives and
it made me a bit curious. Mathematically I was pretty sure I was right,
but it seemed that I got some false positives. I immediately thought
that it could be due to floating point errors and I brought in my z3
solution and checked which collision tests gave different answers for
me. It turns out that it was only a few, but these where all ones with
very close matches.

To make sure this wasn't the problem, I switched to using the
"fractions" library from the python standard library, to get exact
representations of all the formulas. This ended up working wonderfully
and it now gives the correct answer. The problem is that it is really
not as fast as I would want and I want to see if there isn't a solution
that doesn't depend on this.

I have the feeling that I'm going about this the wrong way. There should
be a faster way to do this. I'll check the reddit comments later, maybe
there is some optimization I have missed. For now, I can only think of
the optimization, where we remove a line all together, if it doesn't
intersect the target square at all or in the past. I am not sure there
even are such lines in the input, but I will check!

Another cool thing to note here, is that this is the second time I'm
writing this commit message. I ended up quickly writing this commit, so
that the commit would make it before midnight, but I wanted to rewrite
it to make sure it made sense. I am not sure this is going to work, but
I'll guess we will see in a second!